pyproject.toml
[tool.poetry]
name = "person-service"
version = "0.1.0"
description = "Person service using DDD and Clean Architecture with FastAPI, SQLite and JWT auth"
authors = ["Seu Nome <seu@email.com>"]
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.123.4"
pydantic = "2.12.5"
pydantic-core = "2.41.5"
SQLAlchemy = "2.0.44"
databases = "0.9.0"
aiosqlite = "0.21.0"
httpx = "0.28.1"
python-dotenv = "1.2.1"
email-validator = "2.3.0"
uvicorn = { version = "0.38.0", extras = ["standard"] }
anyio = "4.12.0"
starlette = "0.50.0"
httpcore = "1.0.9"
httptools = "0.7.1"
h11 = "0.16.0"
websockets = "15.0.1"
uvloop = "0.22.1"
watchfiles = "1.1.1"
dnspython = "2.8.0"
typing-extensions = "4.15.0"
typing-inspection = "0.4.2"
annotated-types = "0.7.0"
annotated-doc = "0.0.4"
greenlet = "3.2.4"
certifi = "2025.11.12"
click = "8.3.1"
idna = "3.11"
packaging = "25.0"
Pygments = "2.19.2"
PyYAML = "6.0.3"
python-jose = { version = "3.3.0", extras = ["cryptography"] }
passlib = "1.7.4"
bcrypt = "4.0.1"
python-multipart = "0.0.6"
argon2-cffi = "^21.3"
pep8 = "^1.7.1"
pycodestyle = "^2.14.0"
black = "^25.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "9.0.1"
iniconfig = "2.3.0"
pluggy = "1.6.0"
httpx = "0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"